Complete Professional Dashboard
This file illustrates a well styled and clean color-palleted dashboard along with multiple sheets demonstrating the objectives. I've upgraded my level of building dashboard. This is the new and final version of the dashboard. The dataset is Air_Quality.xls  from the same source data.gov. Took me complete a couple of dedicated hours to achieve this level of designing.